% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rank}
\alias{get_rank}
\title{Get Single Row Guanrank
Calculates the Guanrank for a single row.}
\usage{
get_rank(row, sf, censored, uncensored)
}
\arguments{
\item{row}{vector containing named strings for "event," "n," "time"}

\item{sf}{the survfit object}

\item{censored}{original data frame filtered for censored (no event) samples}

\item{uncensored}{original data frame filtered for uncensored (no event)
samples}
}
\value{
Guanrank (non-normalized) for the row
}
\description{
Get Single Row Guanrank
Calculates the Guanrank for a single row.
}
