% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vectorized_guanrank}
\alias{vectorized_guanrank}
\title{Apply Guanrank to Vectors or Data Frame}
\usage{
vectorized_guanrank(time, event)
}
\arguments{
\item{time}{vector of numeric survival times (possibly censored). Will be
attempted to coerce to numeric via as.numeric}

\item{event}{vector of logicals indicating whether or not an event occurred
(if no event occurred, the sample is considered right-censored)}
}
\value{
vector of numeric Guanrank values, normalized to [0,1], with higher
value indicating higher hazard
}
\description{
Apply Guanrank to Vectors or Data Frame
}
\examples{
time <- c(93, 169, 922, 176, 789, 378, 780, 47, 77, 85, 94, 91, 64, 100)
event <- c(TRUE, TRUE, FALSE, TRUE, FALSE, FALSE, FALSE, TRUE, TRUE, TRUE,
TRUE, TRUE, TRUE, TRUE)
vectorized_guanrank(time, event)

# Additionally, this can be done on columns of a data frame directly
df <- data.frame("pfs" = time, "status" = event)
dplyr::mutate(df, gr = vectorized_guanrank(pfs, status))
}
